# Remove InstrumentedPredicate Locking