**What this PR does**:
Removes locking on the predicate filter. Evaluating the usage in ColumnIterator this seems safe as it's all accessed from the same goroutine.

For larger queries accessing millions or 10s of millions of values this can save 5-7% on query time.

Benchmarks are run against real trace data and evaluating ~30M values:
```
name                             old time/op    new time/op    delta
InstanceSearchCompleteParquet-8     1.77s ± 6%     1.66s ± 5%  -6.25%  (p=0.004 n=9+9)

name                             old alloc/op   new alloc/op   delta
InstanceSearchCompleteParquet-8     320MB ± 1%     320MB ± 1%    ~     (p=0.897 n=10+8)

name                             old allocs/op  new allocs/op  delta
InstanceSearchCompleteParquet-8     2.32M ± 0%     2.32M ± 0%    ~     (p=0.666 n=9+9)
```